Renamed advance_indexing.*pp into integer_advanced_indexing.*pp
Streamlined call operator implementation for projection classes.
Added missing includes.
